move stories to match new file structure
