+++ wls-gui-wahllokalsystem/stories/components/ergebnisermittlung/MBW/stapelBC/BaseKandidatRow.stories.ts
@@ -4,7 +4,7 @@ import { useCommonTestDataFactory } from "@tests/utils/common/CommonTestDataFact
 import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/ergebnisseTestDataFactory.ts";
 import { useWahlvorschlaegeTestDataFactory } from "@tests/utils/wahlvorschlaege/WahlvorschlaegeTestDataFactory.ts";
 
-import BaseKandidatRow from "@/components/ergebnisermittlung/MBW/kandidatenStimmen/BaseKandidatRow.vue";
+import BaseKandidatRow from "@/components/ergebnisermittlung/MBW/stapelBC/BaseKandidatRow.vue";
 
 const meta = {
   component: BaseKandidatRow,
